+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Removed deprecated `get_domain_premium_price`. Use `get_domain_prices` instead.
+- Removed deprecated `get_whois_privacy` (dnsimple/dnsimple-developer#919)
+- Removed deprecated `renew_whois_privacy` (dnsimple/dnsimple-developer#919)
 - Dropped support for OTP 25 (EOL)
 - Dropped support for Elixir 1.14 and 1.15 (EOL)